requirements.txt
python-telegram-bot==21.6
feedparser==6.0.11
aiohttp==3.9.5
